TRQ-3301. Fix an issue with qdel all returning an error.

+++ src/cmds/qdel.c
@@ -363,7 +363,7 @@ cnt:
       prt_job_err((char *)"qdel", connect, job_id_out);
       }
     
-    if (!located && any_failed != 0)
+    if (!located && any_failed != 0 && strcasecmp(job_id, "all"))
       {
       fprintf(stderr, "qdel: nonexistent job id: %s\n", job_id);
       }
